+++ src/GRA.Controllers/ViewModel/MissionControl/Carousels/ListViewModel.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using GRA.Controllers.ViewModel.Shared;
+
+namespace GRA.Controllers.ViewModel.MissionControl.Carousels
+{
+    public class ListViewModel
+    {
+        public IEnumerable<GRA.Domain.Model.Carousel> Carousels { get; set; }
+        public PaginateViewModel PaginateModel { get; set; }
+    }
+}