+++ src/GRA.Controllers/ViewModel/PageAction.cs
@@ -1,0 +1,8 @@
+﻿namespace GRA.Controllers.ViewModel
+{
+    public enum PageAction
+    {
+        New,
+        Edit
+    }
+}